backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kiko-backend"
version = "1.0.0"
description = "Kiko - Japanese Radio Learning Platform Backend API"
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "Kiko Team", email = "team@kiko.app" }
]
keywords = ["japanese", "learning", "radio", "api", "fastapi"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Topic :: Education",
    "Topic :: Multimedia :: Sound/Audio",
]
requires-python = ">=3.11"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "bcrypt==4.3.0",
    "certifi==2025.6.15",
    "cffi==1.17.1",
    "click==8.1.8",
    "cryptography==45.0.4",
    "deprecation==2.1.0",
    "ecdsa==0.19.1",
    "exceptiongroup==1.3.0",
    "fastapi==0.115.14",
    "gotrue==2.12.2",
    "h11==0.16.0",
    "h2==4.2.0",
    "hpack==4.1.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "hyperframe==6.1.0",
    "idna==3.10",
    "iniconfig==2.1.0",
    "packaging==25.0",
    "passlib==1.7.4",
    "pluggy==1.6.0",
    "postgrest==1.1.1",
    "pyasn1==0.6.1",
    "pycparser==2.22",
    "pydantic==2.11.7",
    "pydantic-settings==2.10.1",
    "pydantic_core==2.33.2",
    "Pygments==2.19.2",
    "PyJWT==2.10.1",
    "pytest==8.4.1",
    "pytest-mock==3.14.1",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.1",
    "python-jose==3.5.0",
    "python-multipart==0.0.20",
    "realtime==2.5.3",
    "rsa==4.9.1",
    "six==1.17.0",
    "sniffio==1.3.1",
    "starlette==0.46.2",
    "storage3==0.12.0",
    "StrEnum==0.4.15",
    "supabase==2.16.0",
    "supafunc==0.10.1",
    "tomli==2.2.1",
    "typing-inspection==0.4.1",
    "typing_extensions==4.14.0",
    "uvicorn==0.35.0",
    "websockets==15.0.1",
    "redis==5.0.1",
    "ffmpeg-python==0.2.0",
    "python-magic==0.4.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-mock>=3.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.24.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]

[project.urls]
Homepage = "https://github.com/kiko-team/kiko"
Documentation = "https://docs.kiko.app"
Repository = "https://github.com/kiko-team/kiko.git"
"Bug Tracker" = "https://github.com/kiko-team/kiko/issues"

[project.scripts]
kiko-server = "app.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["app"]

[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true 